🎨 Adapt Footer to be responsive
